create actions module and move game_control under it
